requirements.txt
# requirements.txt - Updated for Gemini support
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0

# Core ML/AI packages
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2

# LangChain packages
langchain==0.1.0
langchain-community==0.0.10
langchain-huggingface==0.0.1

# Google Gemini support
langchain-google-genai==0.0.6
google-generativeai==0.3.2

# Vector DB
faiss-cpu==1.7.4

# HuggingFace
transformers==4.36.0
sentence-transformers==2.2.2
torch==2.1.0

# HTTP client
httpx==0.25.2
requests==2.31.0